+++ sdk/compute/azure-mgmt-compute/generated_tests/test_compute_management_virtual_machine_sizes_operations_async.py
@@ -21,7 +21,7 @@ class TestComputeManagementVirtualMachineSizesOperationsAsync(AzureMgmtRecordedT
 
     @RandomNameResourceGroupPreparer(location=AZURE_LOCATION)
     @recorded_by_proxy_async
-    async def test_list(self, resource_group):
+    async def test_virtual_machine_sizes_list(self, resource_group):
         response = self.client.virtual_machine_sizes.list(
             location="str",
             api_version="2024-07-01",