+++ sdk/ml/azure-ai-ml/azure/ai/ml/operations/_model_operations.py
@@ -109,7 +109,7 @@ class ModelOperations(_ScopeDependentOperations):
         **kwargs,
     ):
         super(ModelOperations, self).__init__(operation_scope, operation_config)
-        ops_logger.update_info(kwargs)
+        ops_logger.update_filter()
         self._model_versions_operation = service_client.model_versions
         self._model_container_operation = service_client.model_containers
         self._service_client = service_client