+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/_quickpulse/_live_metrics.py
@@ -1,14 +1,12 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # cSpell:disable
-from datetime import datetime
-from typing import Any, Iterable
+from typing import Any
 
 import logging
 import platform
 import psutil
 
-from opentelemetry.metrics import CallbackOptions, Observation
 from opentelemetry.sdk._logs import LogData
 from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.resources import Resource
@@ -31,27 +29,40 @@ from azure.monitor.opentelemetry.exporter._quickpulse._constants import (
     _REQUEST_FAILURE_RATE_NAME,
     _REQUEST_RATE_NAME,
 )
+from azure.monitor.opentelemetry.exporter._quickpulse._cpu import (
+    _get_process_memory,
+    _get_process_time_normalized,
+    _get_process_time_normalized_old,
+)
 from azure.monitor.opentelemetry.exporter._quickpulse._exporter import (
     _QuickpulseExporter,
     _QuickpulseMetricReader,
 )
-from azure.monitor.opentelemetry.exporter._quickpulse._generated.models import MonitoringDataPoint
+from azure.monitor.opentelemetry.exporter._quickpulse._filter import (
+    _check_metric_filters,
+)
+from azure.monitor.opentelemetry.exporter._quickpulse._generated.models import (
+    MonitoringDataPoint,
+    TelemetryType,
+)
+from azure.monitor.opentelemetry.exporter._quickpulse._projection import (
+    _create_projections,
+)
 from azure.monitor.opentelemetry.exporter._quickpulse._state import (
     _QuickpulseState,
     _is_post_state,
     _append_quickpulse_document,
     _get_quickpulse_derived_metric_infos,
-    _get_quickpulse_last_process_cpu,
-    _get_quickpulse_last_process_time,
-    _get_quickpulse_process_elapsed_time,
     _set_global_quickpulse_state,
-    _set_quickpulse_last_process_cpu,
-    _set_quickpulse_last_process_time,
-    _set_quickpulse_process_elapsed_time,
 )
-from azure.monitor.opentelemetry.exporter._quickpulse._types import _TelemetryData
+from azure.monitor.opentelemetry.exporter._quickpulse._types import (
+    _DependencyData,
+    _ExceptionData,
+    _RequestData,
+    _TelemetryData,
+    _TraceData,
+)
 from azure.monitor.opentelemetry.exporter._quickpulse._utils import (
-    _derive_metrics_from_telemetry_data,
     _get_log_record_document,
     _get_span_document,
 )
@@ -103,7 +114,7 @@ class _QuickpulseManager(metaclass=Singleton):
         self._base_monitoring_data_point = MonitoringDataPoint(
             version=_get_sdk_version(),
             # Invariant version 5 indicates filtering is supported
-            invariant_version=2,
+            invariant_version=5,
             instance=part_a_fields.get(ContextTagKeys.AI_CLOUD_ROLE_INSTANCE, ""),
             role_name=part_a_fields.get(ContextTagKeys.AI_CLOUD_ROLE, ""),
             machine_name=platform.node(),
@@ -218,43 +229,24 @@ class _QuickpulseManager(metaclass=Singleton):
                 _logger.exception("Exception occurred while recording log record.")
 
 
-# pylint: disable=unused-argument
-def _get_process_memory(options: CallbackOptions) -> Iterable[Observation]:
-    memory = 0
-    try:
-        # rss is non-swapped physical memory a process has used
-        memory = PROCESS.memory_info().rss
-    except (psutil.NoSuchProcess, psutil.AccessDenied):
-        pass
-    yield Observation(memory, {})
+# Filtering
 
-
-# pylint: disable=unused-argument
-def _get_process_time_normalized_old(options: CallbackOptions) -> Iterable[Observation]:
-    normalized_cpu_percentage = 0.0
-    try:
-        cpu_times = PROCESS.cpu_times()
-        # total process time is user + system in s
-        total_time_s = cpu_times.user + cpu_times.system
-        process_time_s = total_time_s - _get_quickpulse_last_process_time()
-        _set_quickpulse_last_process_time(process_time_s)
-        # Find elapsed time in s since last collection
-        current_time = datetime.now()
-        elapsed_time_s = (current_time - _get_quickpulse_process_elapsed_time()).total_seconds()
-        _set_quickpulse_process_elapsed_time(current_time)
-        # Obtain cpu % by dividing by elapsed time
-        cpu_percentage = process_time_s / elapsed_time_s
-        # Normalize by dividing by amount of logical cpus
-        normalized_cpu_percentage = cpu_percentage / NUM_CPUS
-        _set_quickpulse_last_process_cpu(normalized_cpu_percentage)
-    except (psutil.NoSuchProcess, psutil.AccessDenied, ZeroDivisionError):
-        pass
-    yield Observation(normalized_cpu_percentage, {})
-
-
-# pylint: disable=unused-argument
-def _get_process_time_normalized(options: CallbackOptions) -> Iterable[Observation]:
-    yield Observation(_get_quickpulse_last_process_cpu(), {})
-
+# Called by record_span/record_log when processing a span/log_record
+# Derives metrics from projections if applicable to current filters in config
+def _derive_metrics_from_telemetry_data(data: _TelemetryData):
+    metric_infos_dict = _get_quickpulse_derived_metric_infos()
+    metric_infos = []  # type: ignore
+    if isinstance(data, _RequestData):
+        metric_infos = metric_infos_dict.get(TelemetryType.REQUEST)  # type: ignore
+    elif isinstance(data, _DependencyData):
+        metric_infos = metric_infos_dict.get(TelemetryType.DEPENDENCY)  # type: ignore
+    elif isinstance(data, _ExceptionData):
+        metric_infos = metric_infos_dict.get(TelemetryType.EXCEPTION)  # type: ignore
+    elif isinstance(data, _TraceData):
+        metric_infos = metric_infos_dict.get(TelemetryType.TRACE)  # type: ignore
+    if metric_infos and _check_metric_filters(metric_infos, data):
+        # Since this data matches the filter, create projections used to
+        # generate filtered metrics
+        _create_projections(metric_infos, data)
 
 # cSpell:enable