+++ sdk/eventhub/azure-eventhub-checkpointstoreblob-aio/CHANGELOG.md
@@ -11,6 +11,11 @@ This version and all future versions will require Python 3.7+. Python 2.7 and 3.
 - Fixed a bug with `BlobCheckpointStore.claim_ownership` mutating the `ownership_list` argument to no longer mutate the argument.
 - Updated `azure-core` dependecy to 1.20.1 to fix `cchardet` ImportError.
 
+### Other Changes
+
+- Updated vendor azure-storage-blob dependency to v12.24.0.
+  - Fixed typing/linting issues and other bugs. See azure-storage-blob CHANGELOG.md for more info.
+
 ## 1.1.4 (2021-04-07)
 
 This version and all future versions will require Python 2.7 or Python 3.6+, Python 3.5 is no longer supported.